Refactor , move algos to their own folder
