+++ algo/gtree/tree.go
