+++ algo/tree/tree.go
